Move code to test helpers
